fixes to how we construct patches
It would appear (thanks, no documentation!) that the only Authorizer
field that supports patch ops 'add' and 'remove' is providerARNs.  To
accommodate the lack of add and remove, the ignoring of 'default'
values has been stripped to allow for "removing" missing fields.  The
'add' is not really necessary since the Authorizer has a number of
defaulted values and behaviors that aren't represented in the
get_authorizers() output.